requirements.txt
Flask
transformers
pip
setuptools
wheel
bitsandbytes
accelerate
huggingface-hub==0.24.0
torchmetrics
sentence-transformers
llama-index